renamed two-ways relations
